+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/basicPhysicalOperators.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.plans.physical.{Partitioning, RangePartitioning, SinglePartition, UnknownPartitioning}
-import org.apache.spark.sql.execution.{ProjectExec, SampleExec, SparkPlan}
+import org.apache.spark.sql.execution.{FilterExec, ProjectExec, SampleExec, SparkPlan}
 import org.apache.spark.sql.rapids.{GpuPartitionwiseSampledRDD, GpuPoissonSampler}
 import org.apache.spark.sql.rapids.execution.TrampolineUtil
 import org.apache.spark.sql.types.{DataType, LongType}
@@ -777,6 +777,18 @@ object GpuFilter {
       boundCondition: GpuTieredProject): ColumnarBatch = {
     val checkedFilterMask = computeCheckedFilterMask(boundCondition, batch)
     doFilter(checkedFilterMask, batch)
+  }
+}
+
+case class GpuFilterExecMeta(
+  filter: FilterExec,
+  override val conf: RapidsConf,
+  parentMetaOpt: Option[RapidsMeta[_, _, _]],
+  rule: DataFromReplacementRule
+) extends SparkPlanMeta[FilterExec](filter, conf, parentMetaOpt, rule) {
+  override def convertToGpu(): GpuExec = {
+    GpuFilterExec(childExprs.head.convertToGpu(),
+      childPlans.head.convertIfNeeded())(useTieredProject = this.conf.isTieredProjectEnabled)
   }
 }
